aboutme-dl/requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
msal==1.8.0
Office365-REST-Python-Client==2.3.1
openpyxl==3.0.5
pycparser==2.20
PyJWT==1.7.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
